Set exception when attribute isn't found (#97)
I used to think PyObject_GetAttrString() set an exception when returning NULL, but it does not.

So go through and make sure we set an exception where necessary. This mostly effects the aws_py_get_XYZ() functions, which are super copy-pasty, so replace their implementations with helper functions/macros that have great error messages while I'm at it.