More support for sloppy shutdowns.
We were seeing crashes in the logger when a program exited while native objects were still alive. We'd clean up the global static logger while logging was going on and ... crash. Attempt to fix by NOT cleaning up global statics during module unload.

It's not clear to me whether C modules can be reloaded, so this seems like an acceptable solution. If someone show me where the reload process is documented, and has a use case, we can try harder to support it.