More support for sloppy shutdowns. (#165)
We were seeing occasional crashes in the logger code when a program exited while native objects were still alive. The global static logger was cleaned up during module unload, so if another thread is logging in the middle of that ... crash. Attempt to fix by NOT cleaning up global statics during module unload.

It's not clear to me whether C modules can be reloaded, so this seems like an acceptable solution. If someone can show me where the reload process is documented, and has a use case, we can try harder to support it. Maybe track the number of living native objects, and clean up global statics if the count is 0.