Don't run aws-lc's codegen at build time. (#276)
aws-lc uses codegen to create some of its source files, and it runs these generators if the user's machine has Go and Perl installed. However, up-to-date versions of these files are already committed to the repo, so it's not necessary for users to run codegen on their machines.

This should fix the awscrt package on piwheels.org, which compiles wheels for Raspberry Pi. The piwheels build machines have Go installed, but prevent git from making network calls during the build process, and Go uses git to fetch dependencies, and .... man it took a while to figure out what was going wrong there.

Anyway, turn off the unnecessary complexity.