Fix Rasberry Pi stream test failures.
The aws-c libraries were not being compiled with Large Files Support (LFS), which led to confusion over the size of off_t, which led to stack corruption when functions took arguments of this type.

This is fixed in the latest aws-c-io which uses int64_t for offsets instead of off_t/aws_off_t.