First pass at builder simplification for Python V2 SDK
These changes make it where it's not required to create and manage
a DefaultHostResolver, EventLoopGroup, or ClientBootstrap to make
a connection. Instead, default singletons have been added to handle
this usecase.

The constructors of several classes have also been altered to allow
passing a ClientBootstrap to be optional rather than required. When
optional, the ClientBootstrap singleton will be used.

Finally, the singletons were added to the testing to ensure they work
as expected moving forward.