Fix release pipeline (#404)
**Issue:**
A docker image used in the release process started having git errors:
> fatal: detected dubious ownership in repository at '/aws-crt-python'

**Investigation:**
Git 2.35.2+ introduced additional [safety checks](https://github.blog/2022-04-12-git-security-vulnerability-announced/), that raise errors if the `.git/` folder isn't owned by the user running git. The release process runs this docker container while mounting the `aws-crt-python/` directory along with its `.git` folder. Within docker, the user is `root` by default, and when it saw the `.git/` folder owned by someone else, it freaked out.

We haven't seen this error in previous releases because we were accidentally using a very old build of that docker container. When we got a newer build, it had a newer version of git, which raised this error.

**Description of changes:**
- Fix error by running as current user within docker, instead of running as root.
- Unrelated: change "dev" version from `1.0.0-dev` -> `1.0.0.dev0`.  Python didn't like the old version string and normalized it to `1.0.0.dev0` anyway. This gets rid of some warnings at build time.
- Unrelated: remove some lines where we uselessly query version

**Other fixes considered:**
- Considered removing all git calls from the release pipeline, but this would have been a big change.
- Considered modifying all our docker images so they disabled these new safety checks via `git config --global --add safe.directory '*'`, but I wasn't comfortable lowering security like that given my shallow understanding of these issues.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.