Fix bug when S3Request is GC'd before download completes (#506)
**Issue:**
I discovered that, if you drop the reference to an S3Request before the operation completes and it's downloading using `recv_filepath`, then everything looks like it succeeds but you end up with a 0 byte file on disk.

**Description of changes:**
Ensure the file isn't closed until the operation is complete.

Most of our code was careful to keep things alive until the operation was complete. But for whatever reason we put the file-closing code in the wrong place.