Fix build build errors on AL2012, due to linking librt statically (#632)
**Issue:**

Build errors on AL2012, due to linking librt statically


**Diagnosis:**

We have a [tricky bit of code in setup.py](https://github.com/awslabs/aws-crt-python/blob/bc205fa35e28fab19cddd3b9cd506cdc031d98e9/setup.py#L351-L362) where we force SOME libraries to be linked statically. The location of this code is sensitive. Libraries added afterwards are NOT forced to be statically linked, and that's on purpose. That code was moved in a [recent PR](https://github.com/awslabs/aws-crt-python/pull/630), and now it's forcing ALL libraries to be linked statically, even ones that are always available on the OS.

**Description of changes:**

* Move tricky bit of code back where it was
* Remove some old code that only applied to ancient versions of python
* Simplify some code that checks whether or not libcrypto is being used